dom/base/test/test_anonymousContent_style_csp.html

<!DOCTYPE HTML>
<html>
<!-- https://bugzilla.mozilla.org/show_bug.cgi?id=1020244 -->
<meta charset="utf-8">
<title>Test for Bug 1185351 - Make sure that we don't enforce CSP on styles for AnonymousContent</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
<div>
  <div id="test-element" style="color:red;">text content</div>
</div>
<script>
  let chromeDocument = SpecialPowers.wrap(document);
  let testElement = document.querySelector("div");
  let anonymousContent = chromeDocument.insertAnonymousContent();
  anonymousContent.root.appendChild(testElement.cloneNode(true));

  let el = anonymousContent.root.getElementById("test-element");
  el.style = `color: green;`;
  let style = el.getAttribute("style");
  is(style, "color: green;", "The anonymous content exists with CSP");

  chromeDocument.removeAnonymousContent(anonymousContent);
</script>
